+++ vllm/model_executor/layers/quantization/compressed_tensors/schemes/__init__.py
@@ -2,7 +2,7 @@ from .compressed_tensors_scheme import CompressedTensorsScheme
 from .compressed_tensors_unquantized import CompressedTensorsUnquantized
 from .compressed_tensors_w4a16_24 import (CompressedTensorsW4A16Sparse24,
                                           W4A16SPARSE24_SUPPORTED_BITS)
-from .compressed_tensors_w8a8 import CompressedTensorsW8A8
+from .compressed_tensors_w8a8_int8 import CompressedTensorsW8A8Int8
 from .compressed_tensors_w8a8_fp8 import CompressedTensorsW8A8Fp8
 from .compressed_tensors_wNa16 import (CompressedTensorsWNA16,
                                        WNA16_SUPPORTED_BITS)
@@ -12,7 +12,7 @@ __all__ = [
     "CompressedTensorsUnquantized",
     "CompressedTensorsWNA16",
     "CompressedTensorsW4A16Sparse24",
-    "CompressedTensorsW8A8",
+    "CompressedTensorsW8A8Int8",
     "CompressedTensorsW8A8Fp8",
     "WNA16_SUPPORTED_BITS",
     "W4A16SPARSE24_SUPPORTED_BITS",
